Split the settlement worker into config/confirm/settle modules
worker.rs had grown to ~550 lines mixing the run loop, static config, UTXO
confirmation polling, and per-bundle settlement. Promote it to a module root
(mirroring the covenant.rs + covenant/ layout) holding only run() plus the
submodule wiring, and move the rest into:

- worker/config.rs: SettlementMode, SettlementWorkerConfig, AlternationPacer.
- worker/confirm.rs: OutpointAt, ConfirmOutcome, CovenantLiveness, and the
  poll/confirm/liveness helpers plus their poll constants.
- worker/settle.rs: SettleOutcome, settle_one, RejectionClass, classify_rejection.

Cross-submodule internals are pub(super); the public surface (SettlementMode,
SettlementWorkerConfig, run, and the test-utils AlternationPacer) is unchanged,
re-exported through the worker root so lib.rs needs no edit.

Also fix two stale comments in settle_one: repoint a dangling reference from
the never-existent is_retriable_fee_rejection to classify_rejection, and drop a
sim-driver consumer mention.
